+++ test/DePayRouterV1.spec.ts
@@ -1,18 +1,9 @@
 import chai, { expect } from 'chai'
 
-import {
-  Contract,
-  providers,
-  BigNumber,
-  Wallet 
-} from 'ethers'
-
 import { 
   solidity,
-  deployContract,
   deployMockContract,
   loadFixture,
-  MockProvider
 } from 'ethereum-waffle'
 
 import {
@@ -22,8 +13,6 @@ import {
   routerFixture,
   testTokenFixture,
   unapprovedUniswapFixture,
-  uniswapFixture,
-  uniswapPairAndCallContractFixture,
   uniswapPairFixture,
 } from './shared/fixtures'
 
@@ -153,35 +142,35 @@ describe('DePayRouterV1', () => {
   })
 
   it('allows the contract owner to add plugins', async () => {
-    const {router, otherWallet, uniswapPlugin} = await loadFixture(uniswapFixture)
+    const {router, otherWallet, paymentPlugin} = await loadFixture(paymentFixture)
 
-    expect(await router.isApproved(uniswapPlugin.address)).to.eq(true)
+    expect(await router.isApproved(paymentPlugin.address)).to.eq(true)
     expect(await router.isApproved(otherWallet.address)).to.eq(false)
   })
 
   it('emits PluginApproved when approving new plugins', async () => {
-    const {configuration, uniswapPlugin} = await loadFixture(uniswapFixture)
+    const {configuration, paymentPlugin} = await loadFixture(paymentFixture)
 
     await expect(
-      configuration.approvePlugin(uniswapPlugin.address)
+      configuration.approvePlugin(paymentPlugin.address)
     ).to.emit(configuration, 'PluginApproved')
     .withArgs(
-      uniswapPlugin.address
+      paymentPlugin.address
     );
   })
 
   it('does NOT allow others to add plugins', async () => {
-    const {configuration, otherWallet, uniswapPlugin} = await loadFixture(uniswapFixture)
+    const {configuration, otherWallet, paymentPlugin} = await loadFixture(paymentFixture)
 
     await expect(
-      configuration.connect(otherWallet).approvePlugin(uniswapPlugin.address)
+      configuration.connect(otherWallet).approvePlugin(paymentPlugin.address)
     ).to.be.revertedWith(
       'VM Exception while processing transaction: revert Ownable: caller is not the owner'
     )
   })
 
   it('fails when trying to use a plugin that is not approved', async () => {
-    const {router, ownerWallet, otherWallet, uniswapPlugin} = await loadFixture(unapprovedUniswapFixture)
+    const {router, configuration, ownerWallet, otherWallet} = await loadFixture(paymentFixture)
     
     await expect(
       route({
@@ -195,182 +184,6 @@ describe('DePayRouterV1', () => {
       })
     ).to.be.revertedWith(
       'VM Exception while processing transaction: revert DePay: Plugin not approved'
-    )
-  })
-
-  it('swaps tokens via uniswap before performing a payment', async () => {
-    const {
-      otherWallet,
-      ownerWallet,
-      paymentPlugin,
-      router,
-      token0,
-      token1,
-      uniswapPlugin,
-      uniswapRouter,
-      WETH,
-    } = await loadFixture(uniswapPairFixture)
-
-    let path = [token0.address, WETH.address, token1.address]
-    let amountOut = 1000
-    let amounts = await uniswapRouter.getAmountsIn(amountOut, path)
-    let amountIn = amounts[0].toNumber()
-
-    await expect(() => 
-      route({
-        router,
-        wallet: ownerWallet,
-        path: path,
-        amounts: [
-          amountIn,
-          amountOut,
-          now()+10000 // deadline
-        ],
-        addresses: [otherWallet.address],
-        plugins: [uniswapPlugin.address, paymentPlugin.address]
-      })
-    ).to.changeTokenBalance(token1, otherWallet, 1000)
-  })
-
-  it('fails when a miner withholds a swap and executes the payment transaction after the deadline has been reached', async () => {
-    const {
-      otherWallet,
-      ownerWallet,
-      paymentPlugin,
-      router,
-      token0,
-      token1,
-      uniswapPlugin,
-      uniswapRouter,
-      WETH,
-    } = await loadFixture(uniswapPairFixture)
-
-    let path = [token0.address, WETH.address, token1.address]
-    let amountOut = 1000
-    let amounts = await uniswapRouter.getAmountsIn(amountOut, path)
-    let amountIn = amounts[0].toNumber()
-
-    await expect(
-      route({
-        router,
-        wallet: ownerWallet,
-        path: path,
-        amounts: [
-          amountIn,
-          amountOut,
-          now() - 1000 // deadline in the past
-        ],
-        addresses: [otherWallet.address],
-        plugins: [uniswapPlugin.address, paymentPlugin.address]
-      })
-    ).to.be.revertedWith(
-      'UniswapV2Router: EXPIRED'
-    )
-  })
-
-  it('swaps ETH to token via uniswap before performing a payment', async () => {
-    const {
-      otherWallet,
-      ownerWallet,
-      paymentPlugin,
-      router,
-      token0,
-      uniswapPlugin,
-      uniswapRouter,
-      WETH,
-    } = await loadFixture(uniswapPairFixture)
-
-    let amountOut = 1000
-    let amounts = await uniswapRouter.getAmountsIn(amountOut, [WETH.address, token0.address])
-    let amountIn = amounts[0].toNumber()
-
-    await expect(() => 
-      route({
-        router,
-        wallet: ownerWallet,
-        path: [ETH, token0.address],
-        amounts: [
-          amountIn,
-          amountOut,
-          now()+10000 // deadline
-        ],
-        addresses: [otherWallet.address],
-        plugins: [uniswapPlugin.address, paymentPlugin.address],
-        value: amountIn
-      })
-    ).to.changeTokenBalance(token0, otherWallet, amountOut)
-  })
-
-  it('swaps tokens for ETH via uniswap before performing a payment', async () => {
-    const {
-      otherWallet,
-      ownerWallet,
-      paymentPlugin,
-      router,
-      token0,
-      uniswapPlugin,
-      uniswapRouter,
-      WETH,
-    } = await loadFixture(uniswapPairFixture)
-
-    let amountOut = 1000
-    let amounts = await uniswapRouter.getAmountsIn(amountOut, [token0.address, WETH.address])
-    let amountIn = amounts[0].toNumber()
-
-    await expect(() => 
-      route({
-        router,
-        wallet: ownerWallet,
-        path: [token0.address, ETH],
-        amounts: [
-          amountIn,
-          amountOut,
-          now()+10000 // deadline
-        ],
-        addresses: [otherWallet.address],
-        plugins: [uniswapPlugin.address, paymentPlugin.address],
-        value: amountIn
-      })
-    ).to.changeEtherBalance(otherWallet, amountOut)
-  })
-
-  it('swaps tokens to tokens via uniswap and pays the resulting tokens into a contract', async () => {
-    const {
-      contractCallPlugin,
-      exampleContract,
-      ownerWallet,
-      paymentPlugin,
-      router,
-      token0,
-      token1,
-      uniswapPlugin,
-      uniswapRouter,
-      WETH,
-    } = await loadFixture(uniswapPairAndCallContractFixture)
-
-    let path = [token0.address, WETH.address, token1.address]
-    let amountOut = 1000
-    let amounts = await uniswapRouter.getAmountsIn(amountOut, path)
-    let amountIn = amounts[0].toNumber()
-
-    await expect(
-      route({
-        router,
-        wallet: ownerWallet,
-        path: path,
-        amounts: [
-          amountIn,
-          amountOut,
-          now()+10000 // deadline
-        ],
-        addresses: [ownerWallet.address, exampleContract.address],
-        data: ["depositFor(address,uint256)"],
-        plugins: [uniswapPlugin.address, contractCallPlugin.address]
-      })
-    ).to.emit(exampleContract, 'Deposit')
-    .withArgs(
-      ownerWallet.address,
-      amountOut
     )
   })
